pyproject.toml
[tool.poetry]
name = "shredder_challenge"
version = "0.1.0"
description = ""
authors = ["ohtaman <ohtamans@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.0.1"
numpy = "^1.19.4"
PuLP = "^2.3.1"
attrdict = "^2.0.1"
matplotlib = "^3.3.3"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
pylint = "^2.6.0"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
